app/requirements.txt
langchain
huggingface_hub
langchain_community
python-dotenv
langchainhub
fastapi 
uvicorn 
pydantic
fastapi
